[Refactor]: gateway-plugins ext-proc server codebase (#810)
* refactor(gateway-plugins): ext-proc server codebase
