+++ pkg/plugins/gateway/algorithms/least_busy_time.go
@@ -27,6 +27,14 @@ import (
 	"k8s.io/klog/v2"
 )
 
+var (
+	RouterLeastBusyTime Algorithms = "least-busy-time"
+)
+
+func init() {
+	Register(RouterLeastBusyTime, func() (Router, error) { return NewLeastBusyTimeRouter() })
+}
+
 type leastBusyTimeRouter struct {
 	cache *cache.Cache
 }
