+++ pkg/plugins/gateway/algorithms/least_latency.go
@@ -28,6 +28,14 @@ import (
 	"k8s.io/klog/v2"
 )
 
+var (
+	RouterLeastLatency Algorithms = "least-latency"
+)
+
+func init() {
+	Register(RouterLeastLatency, func() (Router, error) { return NewLeastExpectedLatencyRouter() })
+}
+
 type leastExpectedLatencyRouter struct {
 	cache *cache.Cache
 }
