+++ pkg/plugins/gateway/algorithms/least_request.go
@@ -29,6 +29,14 @@ import (
 	"k8s.io/klog/v2"
 )
 
+var (
+	RouterLeastRequest Algorithms = "least-request"
+)
+
+func init() {
+	Register(RouterLeastRequest, func() (Router, error) { return NewLeastRequestRouter() })
+}
+
 type leastRequestRouter struct {
 	cache *cache.Cache
 }
