+++ pkg/plugins/gateway/algorithms/throughput.go
@@ -29,6 +29,14 @@ import (
 	"k8s.io/klog/v2"
 )
 
+var (
+	RouterThroughput Algorithms = "throughput"
+)
+
+func init() {
+	Register(RouterThroughput, func() (Router, error) { return NewThroughputRouter() })
+}
+
 type throughputRouter struct {
 	cache *cache.Cache
 }
